DDF-5495 Provide an extension for Table export download behavior
