+++ catalog/ui/catalog-ui-search/pom.xml
@@ -463,7 +463,10 @@
                             *
                         </Import-Package>
                         <Export-Package>
-                            org.codice.ddf.catalog.ui.metacard.internal
+                            org.codice.ddf.catalog.ui.metacard.internal,
+                            org.codice.ddf.catalog.ui.api,
+                            org.codice.ddf.catalog.ui.query.cql,
+                            org.codice.ddf.catalog.ui.metacard.transform,
                         </Export-Package>
                         <Karaf-Commands>org.codice.ddf.catalog.ui.forms.commands.*</Karaf-Commands>
                     </instructions>
