+++ catalog/ui/catalog-ui-search/src/main/java/org/codice/ddf/catalog/ui/api/EndpointUtility.java
@@ -1,0 +1,13 @@
+package org.codice.ddf.catalog.ui.api;
+
+import ddf.catalog.federation.FederationException;
+import ddf.catalog.source.SourceUnavailableException;
+import ddf.catalog.source.UnsupportedQueryException;
+import org.codice.ddf.catalog.ui.query.cql.CqlQueryResponse;
+import org.codice.ddf.catalog.ui.query.cql.CqlRequest;
+
+public interface EndpointUtility {
+
+  CqlQueryResponse executeCqlQuery(CqlRequest cqlRequest)
+      throws UnsupportedQueryException, SourceUnavailableException, FederationException;
+}
