+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/table-export/container.tsx
@@ -14,111 +14,19 @@
  **/
 import * as React from 'react'
 import TableExportComponent from './presentation'
-import { exportResultSet } from '../utils/export'
 import LoadingCompanion from '../loading-companion'
-import saveFile from '../utils/save-file'
 import { hot } from 'react-hot-loader'
-const _ = require('underscore')
-const user = require('../../component/singletons/user-instance.js')
 const properties = require('../../js/properties.js')
-const announcement = require('../../component/announcement/index.jsx')
-const Sources = require('../../component/singletons/sources-instance.js')
-const contentDisposition = require('content-disposition')
-function buildCqlQueryFromMetacards(metacards: any) {
-  const queryParts = metacards.map((metacard: any) => {
-    return `(("id" ILIKE '${metacard.metacard.id}'))`
-  })
-  return `(${queryParts.join(' OR ')})`
-}
-const visibleData = (selectionInterface: any) =>
-  buildCqlQueryFromMetacards(
-    selectionInterface.getActiveSearchResults().toJSON()
-  )
-const allData = (selectionInterface: any) =>
-  selectionInterface.getCurrentQuery().get('cql')
-function getCqlForSize(exportSize: string, selectionInterface: any) {
-  return exportSize === 'visible'
-    ? visibleData(selectionInterface)
-    : allData(selectionInterface)
-}
-function getSrcs(selectionInterface: any) {
-  const srcs = selectionInterface.getCurrentQuery().get('src')
-  return srcs === undefined ? _.pluck(Sources.toJSON(), 'id') : srcs
-}
-function getColumnOrder(): string[] {
-  return user
-    .get('user')
-    .get('preferences')
-    .get('columnOrder')
-}
-function getHiddenFields(): string[] {
-  return user
-    .get('user')
-    .get('preferences')
-    .get('columnHide')
-}
-function getHits(sources: Source[]): number {
-  return sources
-    .filter(source => source.id !== 'cache')
-    .reduce((hits, source) => (source.hits ? hits + source.hits : hits), 0)
-}
-function getExportCount({
-  exportSize,
-  selectionInterface,
-  customExportCount,
-}: ExportCountInfo): number {
-  if (exportSize === 'custom') {
-    return customExportCount
-  }
-  const result = selectionInterface.getCurrentQuery().get('result')
-  return exportSize === 'all'
-    ? getHits(result.get('status').toJSON())
-    : result.get('results').length
-}
-function getSorts(selectionInterface: any) {
-  return selectionInterface.getCurrentQuery().get('sorts')
-}
-function getWarning(exportCountInfo: ExportCountInfo): string {
-  const exportCount = getExportCount(exportCountInfo)
-  const result = exportCountInfo.selectionInterface
-    .getCurrentQuery()
-    .get('result')
-  const totalHits = getHits(result.get('status').toJSON())
-  const limitWarning = `You cannot export more than the administrator configured limit of ${
-    properties.exportResultLimit
-  }.`
-  let warningMessage = ''
-  if (exportCount > properties.exportResultLimit) {
-    if (exportCountInfo.exportSize === 'custom') {
-      return limitWarning
-    }
-    warningMessage =
-      limitWarning +
-      `  Only ${properties.exportResultLimit} ${
-        properties.exportResultLimit === 1 ? `result` : `results`
-      } will be exported.`
-  }
-  if (exportCountInfo.exportSize === 'custom') {
-    if (exportCount > totalHits) {
-      warningMessage = `You are trying to export ${exportCount} results but there ${
-        totalHits === 1 ? `is` : `are`
-      } only ${totalHits}.  Only ${totalHits} ${
-        totalHits === 1 ? `result` : `results`
-      } will be exported.`
-    }
-  }
-  if (
-    totalHits > 100 &&
-    exportCount > 100 &&
-    properties.exportResultLimit > 100
-  ) {
-    warningMessage += `  This may take a long time.`
-  }
-  return warningMessage
-}
+import {
+  ExportCountInfo,
+  DownloadInfo,
+} from '../../react-component/utils/export'
 type Props = {
   selectionInterface: () => void
   exportFormats: Option[]
+  getWarning: (exportCountInfo: ExportCountInfo) => string
+  onDownloadClick: (downloadInfo: DownloadInfo) => void
+  filteredAttributes: string[]
 }
 type Option = {
   label: string
@@ -128,15 +36,6 @@ type State = {
   exportSizes: Option[]
   exportFormat: string
   exportSize: string
-  customExportCount: number
-}
-type Source = {
-  id: string
-  hits: number
-}
-interface ExportCountInfo {
-  exportSize: string
-  selectionInterface: any
   customExportCount: number
 }
 export default hot(module)(
@@ -177,54 +76,6 @@ export default hot(module)(
     handleCustomExportCountChange = (value: number) => {
       this.setState({ customExportCount: value })
     }
-    onDownloadClick = async () => {
-      const exportFormat = encodeURIComponent(this.state.exportFormat)
-      const { exportSize, customExportCount } = this.state
-      const { selectionInterface } = this.props
-      try {
-        const hiddenFields = getHiddenFields()
-        const columnOrder = getColumnOrder()
-        const cql = getCqlForSize(exportSize, selectionInterface)
-        const srcs = getSrcs(selectionInterface)
-        const sorts = getSorts(selectionInterface)
-        const count = Math.min(
-          getExportCount({ exportSize, selectionInterface, customExportCount }),
-          properties.exportResultLimit
-        )
-        const args = {
-          hiddenFields: hiddenFields.length > 0 ? hiddenFields : [],
-          columnOrder: columnOrder.length > 0 ? columnOrder : {},
-          columnAliasMap: properties.attributeAliases,
-        }
-        const body = {
-          cql,
-          srcs,
-          count,
-          sorts,
-          args,
-        }
-        const response = await exportResultSet(exportFormat, body)
-        this.onDownloadSuccess(response)
-      } catch (error) {
-        console.error(error)
-      }
-    }
-    async onDownloadSuccess(response: Response) {
-      if (response.status === 200) {
-        const data = await response.blob()
-        const contentType = response.headers.get('content-type')
-        const filename = contentDisposition.parse(
-          response.headers.get('content-disposition')
-        ).parameters.filename
-        saveFile(filename, 'data:' + contentType, data)
-      } else {
-        announcement.announce({
-          title: 'Error',
-          message: 'Could not export results.',
-          type: 'error',
-        })
-      }
-    }
     render() {
       const {
         exportFormat,
@@ -232,7 +83,13 @@ export default hot(module)(
         exportSize,
         customExportCount,
       } = this.state
-      const { exportFormats, selectionInterface } = this.props
+      const {
+        exportFormats,
+        selectionInterface,
+        onDownloadClick,
+        getWarning,
+        filteredAttributes,
+      } = this.props
       return (
         <LoadingCompanion loading={exportFormats.length === 0}>
           {exportFormats.length > 0 ? (
@@ -244,7 +101,15 @@ export default hot(module)(
               handleExportFormatChange={this.handleExportFormatChange}
               handleExportSizeChange={this.handleExportSizeChange}
               handleCustomExportCountChange={this.handleCustomExportCountChange}
-              onDownloadClick={this.onDownloadClick}
+              onDownloadClick={() =>
+                onDownloadClick({
+                  exportFormat,
+                  exportSize,
+                  selectionInterface,
+                  customExportCount,
+                  filteredAttributes,
+                })
+              }
               warning={getWarning({
                 exportSize,
                 selectionInterface,
