+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/utils/export/index.tsx
@@ -18,4 +18,6 @@ export {
   exportResultSet,
   Transformer,
   ResultSet,
+  ExportCountInfo,
+  DownloadInfo,
 } from './export'
